requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

SQLAlchemy~=2.0.43
starlette~=0.27.0
pydantic~=2.11.9
passlib~=1.7.4